requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
Flask-Migrate>=4.0
Flask-Cors>=3.0
python-dotenv>=1.0
flask-restful>=0.3.9
flask-swagger-ui>=3.36.0
marshmallow>=3.20
flasgger==0.9.7.1
reportlab
